refactor: logout functionality
