+++ src/components/Navigation/navigation.ts
@@ -1,14 +1,4 @@
-import Cookies from 'js-cookie';
-import { TOKEN_COOKIE_NAME, removeAuthToken } from '../../lib/jwt';
-import { REDIRECT_PAGE_AFTER_AUTH } from '../Authenticator/authenticator.ts';
-
-export function logout() {
-  localStorage.removeItem(REDIRECT_PAGE_AFTER_AUTH);
-  removeAuthToken();
-
-  // Reloading will automatically redirect the user if required
-  window.location.href = '/';
-}
+import { logout } from '../../lib/auth';
 
 function bindEvents() {
   document.addEventListener('click', (e) => {
